feat: Manage Connection objects with Lua's GC
We will be exposing Connection objects directly to Lua, so it only makes
sense that their memory be managed by Lua.

When an incoming connection is detected, a new userdata containing a
Connection object is created, and added to the "connections" table in
the registry. When the client closes the connection, the uv_close_cb
sets m_connected to false, then removes the Connection object from this table.

m_connected was introduced in case someone else holds a strong reference
to the Connection, preventing it from being automatically garbage
collected, it will be exposed to Lua.