feat: Create client handler coroutine, run client handler
Each connection has an associated Lua coroutine / thread, on which the
client handler runs. It is associated with each Connection object via
the first Lua uservalue of the Connection userdata, though this is
subject to change.

The thread currently shares the global environment, however I plan to
set a custom _ENV for each Connection thread, with a table that provides
a direct reference to the connection / player, plus helper functions to
interact with the user. Access to _G will still be provided, however, by
__index and __newindex metamethods.